refactor(aetherd): 2.4 — split generic message enums out of vendor CommandParser.h (first EB3 decouple)
Closes #4082 (with a corrected target — see below). The MessageType /
MessageSeverity enums were defined in core/CommandParser.h, which the
touchpoint audit tags vendor(flex) because it parses the SmartSDR wire. Every
above-seam file that only needs the generic severity/type enum was forced to
include the vendor parser and trip EB3.

Moves the two enums into a new vendor-free header core/RadioMessageTypes.h;
CommandParser.h re-includes it, so ParsedMessage and all existing includers
compile unchanged. Above-seam consumers of just the enums now include the
neutral header and drop the vendor coupling:

- src/gui/MainWindow.h  — MessageSeverity only → RadioMessageTypes.h
- src/gui/MainWindow.cpp — MessageSeverity only → RadioMessageTypes.h

EB3 baseline: CommandParser removed from both MainWindow rows (7 → 6 each);
the checker reconciles at 76 EB3-known / 0 blocking.

Corrected target vs the issue: #4082 predicted RadioModel.h would drop to 8.
Verification showed RadioModel.h also consumes ParsedMessage (the onMessage
Received(const ParsedMessage&) slot at RadioModel.h:600) — a genuine SmartSDR
wire DTO, not just the enum. Relabeling ParsedMessage as vendor-free to force
the number down would game the ratchet, so RadioModel.h correctly stays coupled
until onMessageReceived is migrated behind the backend (a deeper seam step).
The enum split instead frees MainWindow.h + MainWindow.cpp — two files, the
honest set. RadioModel.cpp and TransmitInhibitPolicy.h likewise keep the vendor
include (they use CommandParser / ParsedMessage).

Behavior-neutral: enums unchanged (values, namespace, call sites); build clean;
71/71 ctest green; engine-boundary --strict clean.
