+++ tools/check_engine_boundary.py
@@ -159,8 +159,8 @@ KNOWN_VENDOR_INCLUDE_BASELINE = {
     "src/gui/DvkPanel.cpp": ["DvkWavTransfer"],
     "src/gui/KiwiPublicReceiverPicker.h": ["KiwiPublicDirectory"],
     "src/gui/KiwiSdrApplet.h": ["KiwiSdrClient"],
-    "src/gui/MainWindow.cpp": ["CommandParser", "DvkWavTransfer", "KiwiSdrManager", "PanadapterStream", "RadioStatusOwnership", "StreamStatus", "TunerModel"],
-    "src/gui/MainWindow.h": ["AntennaGeniusModel", "CommandParser", "FlexControlManager", "PgxlConnection", "SmartLinkClient", "TgxlConnection", "WanConnection"],
+    "src/gui/MainWindow.cpp": ["DvkWavTransfer", "KiwiSdrManager", "PanadapterStream", "RadioStatusOwnership", "StreamStatus", "TunerModel"],
+    "src/gui/MainWindow.h": ["AntennaGeniusModel", "FlexControlManager", "PgxlConnection", "SmartLinkClient", "TgxlConnection", "WanConnection"],
     "src/gui/MainWindowHelpers.cpp": ["PanadapterStream", "SmartLinkClient"],
     "src/gui/MainWindow_Controllers.cpp": ["KiwiSdrProtocol"],
     "src/gui/MainWindow_KiwiSdr.cpp": ["KiwiSdrClient", "KiwiSdrManager", "KiwiSdrProtocol"],
